Modified Generate and Send OTP
