+++ src/main/java/com/pst/login/configuration/LoginConfiguration.java
@@ -1,0 +1,13 @@
+package com.pst.login.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class LoginConfiguration {
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
