--- src/main/java/com/pst/login/repository/LoginRepository.java
@@ -1,10 +1,0 @@
-package com.pst.login.repository;
-
-import com.pst.login.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface LoginRepository extends JpaRepository<UserEntity, Long> {
-
-}
