+++ src/main/java/com/pst/login/request/EmailRequest.java
@@ -1,37 +1,14 @@
 package com.pst.login.request;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
 public class EmailRequest {
     private String toEmail;
     private String subject;
     private String body;
-
-    public String getToEmail() {
-        return toEmail;
-    }
-
-    public void setToEmail(String toEmail) {
-        this.toEmail = toEmail;
-    }
-
-    public String getSubject() {
-        return subject;
-    }
-
-    public void setSubject(String subject) {
-        this.subject = subject;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
-    }
-
-    public EmailRequest(String toEmail, String subject, String body) {
-        this.toEmail = toEmail;
-        this.subject = subject;
-        this.body = body;
-    }
 }
