+++ src/main/java/com/pst/login/service/LoginService.java
@@ -1,10 +1,8 @@
 package com.pst.login.service;
 
-import com.pst.login.entity.UserEntity;
-import com.pst.login.repository.LoginRepository;
 import com.pst.login.request.EmailRequest;
+import com.pst.login.request.UserRequest;
 import com.pst.login.response.UserResponse;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -17,10 +15,6 @@ public class LoginService {
 
     @Autowired
     private RestTemplate restTemplate;
-
-    @Autowired
-    private LoginRepository loginRepository;
-
 
     @Value("${user_service_url}")
     private String userServiceUrl;
@@ -38,16 +32,18 @@ public class LoginService {
 
         int otp = 100000 + new Random().nextInt(900000);
 
-        UserEntity userEntity = new UserEntity(userResponse);
-        userEntity.setOtp(otp);
-        loginRepository.save(userEntity);
+        UserRequest userRequest = new UserRequest(userResponse);
+        userRequest.setOtp(otp);
 
+
+        restTemplate.put(userServiceUrl + "/update-user/" + aadhaarNumber, userRequest, UserResponse.class);
 
         EmailRequest emailRequest = new EmailRequest(userResponse.getEmail(), "Your OTP Code", "Your OTP is: " + otp);
         String messageServiceResponse = restTemplate.postForObject(messageServiceUrl, emailRequest, String.class);
 
 
         return (messageServiceResponse == null) ? "Failed to send OTP" : "OTP sent to: " + userResponse.getEmail();
+
     }
 
 }
